QuicｋHullのテストプログラムの追加． QuicｋHull対策．
接触点が複数で場所がすべて同一の場合に，CreateFirstPlaneで問題がおきていた．
２D版は対策をしたが，3D版はまだ．


git-svn-id: http://springhead.info/spr2/Springhead2/trunk@1608 05cee5c3-a2e9-0310-9523-9dfc2f93dbe1

+++ src/Collision/CDQuickHull3DImp.h
@@ -285,12 +285,12 @@ void CDQHPlanes<TVtx>::CreateCone(TVtx* firstVtx, TVtx* top){
 			for(TVtx** p = vtxBeginInput; p != vtxEndInput; ++p){
 				DSTR << (*p)->GetPos() << std::endl;
 			}
-			DSTR << vtxEndInput - vtxBeginInput << " vertices." << std::endl;
+			DSTR << (int)(vtxEndInput - vtxBeginInput) << " vertices." << std::endl;
 
 			curHorizon = firstVtx->horizon;
 			for(TVtx** p = vtxBeginInput; p != vtxEndInput; ++p){
 				if (firstVtx == *p){
-					DSTR << p - vtxBeginInput << " ";
+					DSTR << (int)(p - vtxBeginInput) << " ";
 				}
 			}
 			for(int i=0; i<len; ++i){
@@ -300,7 +300,7 @@ void CDQHPlanes<TVtx>::CreateCone(TVtx* firstVtx, TVtx* top){
 				curHorizon = curVtx->horizon;
 				for(TVtx** p = vtxBeginInput; p != vtxEndInput; ++p){
 					if (curVtx == *p){
-						DSTR << p - vtxBeginInput << " ";
+						DSTR << int(p - vtxBeginInput) << " ";
 					}
 				}
 			}
